% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_atacseq_data}
\alias{human_atacseq_data}
\title{Human ATAC-seq data}
\format{
An object of class \code{dgCMatrix} with 43 rows and 9834 columns.
}
\usage{
human_atacseq_data
}
\description{
Human ATAC-seq data
}
\keyword{datasets}
