% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_cell_type}
\alias{human_cell_type}
\title{Human cell type information}
\format{
An object of class \code{data.frame} with 9834 rows and 2 columns.
}
\usage{
human_cell_type
}
\description{
Human cell type information
}
\keyword{datasets}
