% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_meta_data}
\alias{human_meta_data}
\title{Human metadata}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9834 rows and 3 columns.
}
\usage{
human_meta_data
}
\description{
Human metadata
}
\keyword{datasets}
