% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_peaks_gr}
\alias{human_peaks_gr}
\title{Human peaks GRanges}
\format{
An object of class \code{GRanges} of length 108377.
}
\usage{
human_peaks_gr
}
\description{
Human peaks GRanges
}
\keyword{datasets}
