% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_rnaseq_data}
\alias{human_rnaseq_data}
\title{Human RNA-seq data}
\format{
An object of class \code{dgCMatrix} with 36601 rows and 9834 columns.
}
\usage{
human_rnaseq_data
}
\description{
Human RNA-seq data
}
\keyword{datasets}
