% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_atacseq_data}
\alias{mouse_atacseq_data}
\title{Mouse ATAC-seq data}
\format{
An object of class \code{data.frame} with 23 rows and 85 columns.
}
\usage{
mouse_atacseq_data
}
\description{
Mouse ATAC-seq data
}
\keyword{datasets}
