% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_peaks_gr}
\alias{mouse_peaks_gr}
\title{Mouse peaks GRanges}
\format{
An object of class \code{GRanges} of length 512595.
}
\usage{
mouse_peaks_gr
}
\description{
Mouse peaks GRanges
}
\keyword{datasets}
