% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_rnaseq_data}
\alias{mouse_rnaseq_data}
\title{Mouse RNA-seq data}
\format{
An object of class \code{data.frame} with 2 rows and 85 columns.
}
\usage{
mouse_rnaseq_data
}
\description{
Mouse RNA-seq data
}
\keyword{datasets}
